feat: undo some changes, move the configmap to point to an env var for default value
