+++ .github/workflows/deploy.yml
@@ -38,6 +38,8 @@ on:
         required: true
       KUBE_CERT:
         required: true
+      BYPASS_PASSWORD:
+        required: true
 
 jobs:
   deploy:
@@ -95,6 +97,7 @@ jobs:
             --set ingress.host=${KUBE_NAMESPACE}.apps.live.cloud-platform.service.justice.gov.uk \
             --set "ipAllowlist=${{ steps.ip-allowlist.outputs.cidrs }}" \
             --set sentryEnvironment=${SENTRY_ENV} \
+            --set maintenance.bypassPassword=${BYPASS_PASSWORD} \
             --wait \
             --timeout 5m
         env:
